Clerbie/requirements.txt
anyio==4.7.0
asgiref==3.8.1
certifi==2024.8.30
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.4
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msgpack==1.1.0
openapi-codec==1.3.2
packaging==24.2
phonenumbers==8.13.52
pillow==11.0.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydot==3.0.4
pydotplus==2.0.2
PyJWT==2.10.1
pyparsing==3.2.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
simplejson==3.19.3
sniffio==1.3.1
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
